impl insert method and added print function
